Move undulator and xbpm tests
