+++ tests/unit_tests/hyperion/experiment_plans/test_hyperion_flyscan_xray_centre_plan.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from unittest.mock import ANY, MagicMock, call, patch
+from unittest.mock import MagicMock, call, patch
 
 import pytest
 from bluesky.run_engine import RunEngine
@@ -30,10 +30,6 @@ from mx_bluesky.common.external_interaction.callbacks.xray_centre.nexus_callback
 )
 from mx_bluesky.common.parameters.constants import (
     DeviceSettingsConstants,
-    PlanNameConstants,
-)
-from mx_bluesky.hyperion.experiment_plans.grid_detect_then_xray_centre_plan import (
-    _gridscan_with_undulator_checks,
 )
 from mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan import (
     SmargonSpeedException,
@@ -43,7 +39,9 @@ from mx_bluesky.hyperion.external_interaction.config_server import HyperionFeatu
 from mx_bluesky.hyperion.parameters.device_composites import (
     HyperionFlyScanXRayCentreComposite,
 )
-from mx_bluesky.hyperion.parameters.gridscan import HyperionSpecifiedThreeDGridScan
+from mx_bluesky.hyperion.parameters.gridscan import (
+    HyperionSpecifiedThreeDGridScan,
+)
 from tests.conftest import (
     RunEngineSimulator,
 )
@@ -288,113 +286,3 @@ class TestFlyscanXrayCentrePlan:
         msgs = assert_message_and_return_remaining(
             msgs, lambda msg: msg.command == "unstage" and msg.obj.name == "panda"
         )
-
-    @patch(
-        "mx_bluesky.common.preprocessors.preprocessors.check_and_pause_feedback",
-        autospec=True,
-    )
-    @patch(
-        "mx_bluesky.common.preprocessors.preprocessors.unpause_xbpm_feedback_and_set_transmission_to_1",
-        autospec=True,
-    )
-    @patch(
-        "mx_bluesky.common.experiment_plans.common_flyscan_xray_centre_plan.run_gridscan",
-    )
-    def test_flyscan_xray_centre_unpauses_xbpm_feedback_on_exception(
-        self,
-        fake_run_gridscan: MagicMock,
-        mock_unpause_and_set_transmission: MagicMock,
-        mock_check_and_pause: MagicMock,
-        hyperion_flyscan_xrc_composite: HyperionFlyScanXRayCentreComposite,
-        hyperion_fgs_params: HyperionSpecifiedThreeDGridScan,
-        beamline_specific: BeamlineSpecificFGSFeatures,
-        RE: RunEngine,
-    ):
-        fake_run_gridscan.side_effect = Exception
-        with pytest.raises(Exception):  # noqa: B017
-            RE(
-                _gridscan_with_undulator_checks(
-                    hyperion_flyscan_xrc_composite,
-                    hyperion_fgs_params,
-                    beamline_specific,
-                )
-            )
-
-        # Called once on exception and once on close_run
-        mock_unpause_and_set_transmission.assert_has_calls([call(ANY, ANY)])
-
-    @patch(
-        "mx_bluesky.hyperion.experiment_plans.hyperion_flyscan_xray_centre_plan.bps.wait"
-    )
-    @patch(
-        "mx_bluesky.common.experiment_plans.inner_plans.do_fgs.check_topup_and_wait_if_necessary",
-    )
-    def test_flyscan_xray_centre_pauses_and_unpauses_xbpm_feedback_in_correct_order(
-        self,
-        mock_check_topup,
-        mock_wait,
-        sim_run_engine: RunEngineSimulator,
-        hyperion_fgs_params: HyperionSpecifiedThreeDGridScan,
-        hyperion_flyscan_xrc_composite: HyperionFlyScanXRayCentreComposite,
-        beamline_specific: BeamlineSpecificFGSFeatures,
-    ):
-        simulate_xrc_result(
-            sim_run_engine, hyperion_flyscan_xrc_composite.zocalo, TEST_RESULT_LARGE
-        )
-
-        msgs = sim_run_engine.simulate_plan(
-            _gridscan_with_undulator_checks(
-                hyperion_flyscan_xrc_composite, hyperion_fgs_params, beamline_specific
-            )
-        )
-
-        # Assert order: pause -> open run -> close run -> unpause (set attenuator)
-        msgs = assert_message_and_return_remaining(
-            msgs,
-            lambda msg: msg.command == "trigger" and msg.obj.name == "xbpm_feedback",
-        )
-        msgs = assert_message_and_return_remaining(
-            msgs,
-            lambda msg: msg.command == "open_run"
-            and msg.run == PlanNameConstants.GRIDSCAN_OUTER,
-        )
-
-        msgs = assert_message_and_return_remaining(
-            msgs,
-            lambda msg: msg.command == "close_run"
-            and msg.run == PlanNameConstants.GRIDSCAN_OUTER,
-        )
-
-        msgs = assert_message_and_return_remaining(
-            msgs,
-            lambda msg: msg.command == "set"
-            and msg.obj.name == "attenuator"
-            and msg.args == (1.0,),
-        )
-
-    @patch(
-        "mx_bluesky.common.experiment_plans.common_flyscan_xray_centre_plan.run_gridscan",
-    )
-    @patch(
-        "dodal.plans.preprocessors.verify_undulator_gap.verify_undulator_gap",
-    )
-    def test_flyscan_xray_centre_does_undulator_check_before_collection(
-        self,
-        mock_verify_gap: MagicMock,
-        mock_plan: MagicMock,
-        RE: RunEngine,
-        hyperion_fgs_params: HyperionSpecifiedThreeDGridScan,
-        hyperion_flyscan_xrc_composite: HyperionFlyScanXRayCentreComposite,
-        beamline_specific: BeamlineSpecificFGSFeatures,
-    ):
-        mock_plan.side_effect = CompleteException
-        with pytest.raises(CompleteException):
-            RE(
-                _gridscan_with_undulator_checks(
-                    hyperion_flyscan_xrc_composite,
-                    hyperion_fgs_params,
-                    beamline_specific,
-                )
-            )
-
-        mock_verify_gap.assert_called_once()
